run.py
import os
from dotenv import load_dotenv

# Load .env FIRST before anything else reads env vars
load_dotenv()

from app import create_app

config_name = os.environ.get('FLASK_ENV', 'development')
app = create_app(config_name)

if __name__ == '__main__':
    app.run(debug=app.config['DEBUG'], host='0.0.0.0', port=5000)
