requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
python-dotenv==1.0.0
Werkzeug==2.3.6
requests==2.31.0
PyMySQL==1.1.0
beautifulsoup4==4.12.3
